re-arrange SecureRandom native part to follow JRuby naming
