+++ core/src/main/java/org/jruby/ext/securerandom/RubySecureRandom.java
@@ -1,0 +1,84 @@
+/*
+ ***** BEGIN LICENSE BLOCK *****
+ * Version: EPL 1.0/GPL 2.0/LGPL 2.1
+ *
+ * The contents of this file are subject to the Eclipse Public
+ * License Version 1.0 (the "License"); you may not use this file
+ * except in compliance with the License. You may obtain a copy of
+ * the License at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Software distributed under the License is distributed on an "AS
+ * IS" basis, WITHOUT WARRANTY OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * rights and limitations under the License.
+ *
+ * Copyright (C) 2017 JRuby Community
+ *
+ * Alternatively, the contents of this file may be used under the terms of
+ * either of the GNU General Public License Version 2 or later (the "GPL"),
+ * or the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
+ * in which case the provisions of the GPL or the LGPL are applicable instead
+ * of those above. If you wish to allow use of your version of this file only
+ * under the terms of either the GPL or the LGPL, and not to allow others to
+ * use your version of this file under the terms of the EPL, indicate your
+ * decision by deleting the provisions above and replace them with the notice
+ * and other provisions required by the GPL or the LGPL. If you do not delete
+ * the provisions above, a recipient may use your version of this file under
+ * the terms of any one of the EPL, the GPL or the LGPL.
+ ***** END LICENSE BLOCK *****/
+package org.jruby.ext.securerandom;
+
+import org.jruby.RubyString;
+import org.jruby.anno.JRubyMethod;
+import org.jruby.anno.JRubyModule;
+import org.jruby.runtime.ThreadContext;
+import org.jruby.runtime.builtin.IRubyObject;
+import org.jruby.util.ConvertBytes;
+
+/**
+ * securerandom.rb native parts.
+ */
+@JRubyModule(name = "SecureRandom")
+public class RubySecureRandom {
+
+    @JRubyMethod(meta = true)
+    public static IRubyObject random_bytes(ThreadContext context, IRubyObject self) {
+        return RubyString.newStringNoCopy(context.runtime, nextBytes(context, 16));
+    }
+
+    @JRubyMethod(meta = true)
+    public static IRubyObject random_bytes(ThreadContext context, IRubyObject self, IRubyObject n) {
+        return RubyString.newStringNoCopy(context.runtime, nextBytes(context, n));
+    }
+
+    @JRubyMethod(meta = true)
+    public static IRubyObject hex(ThreadContext context, IRubyObject self) {
+        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.twosComplementToHexBytes(nextBytes(context, 16), false));
+    }
+
+    @JRubyMethod(meta = true)
+    public static IRubyObject hex(ThreadContext context, IRubyObject self, IRubyObject n) {
+        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.twosComplementToHexBytes(nextBytes(context, n), false));
+    }
+
+    @JRubyMethod(meta = true)
+    public static IRubyObject uuid(ThreadContext context, IRubyObject self) {
+        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.bytesToUUIDBytes(nextBytes(context, 16), false));
+    }
+
+    private static byte[] nextBytes(ThreadContext context, IRubyObject n) {
+        int size = n.isNil() ? 16 : (int) n.convertToInteger().getLongValue();
+
+        return nextBytes(context, size);
+    }
+
+    private static byte[] nextBytes(ThreadContext context, int size) {
+        if (size < 0) throw context.runtime.newArgumentError("negative argument: " + size);
+
+        byte[] bytes = new byte[size];
+        context.getSecureRandom().nextBytes(bytes);
+
+        return bytes;
+    }
+
+}
