+++ core/src/main/java/org/jruby/ext/securerandom/SecureRandomLibrary.java
@@ -1,52 +1,48 @@
 package org.jruby.ext.securerandom;
 
-import org.jruby.RubyString;
-import org.jruby.anno.JRubyMethod;
+import org.jruby.Ruby;
+import org.jruby.RubyModule;
 import org.jruby.runtime.ThreadContext;
 import org.jruby.runtime.builtin.IRubyObject;
-import org.jruby.util.ConvertBytes;
+import org.jruby.runtime.load.Library;
 
 /**
  * Created by headius on 1/14/14.
  */
-public class SecureRandomLibrary {
-    @JRubyMethod(meta = true)
+public class SecureRandomLibrary implements Library {
+
+    public static void load(Ruby runtime) {
+        RubyModule SecureRandom = runtime.defineModule("SecureRandom");
+        SecureRandom.defineAnnotatedMethods(RubySecureRandom.class);
+    }
+
+    public void load(Ruby runtime, boolean wrap) {
+        SecureRandomLibrary.load(runtime);
+    }
+
+    @Deprecated
     public static IRubyObject random_bytes(ThreadContext context, IRubyObject self) {
-        return RubyString.newStringNoCopy(context.runtime, nextBytes(context, 16));
+        return RubySecureRandom.random_bytes(context, self);
     }
 
-    @JRubyMethod(meta = true)
+    @Deprecated
     public static IRubyObject random_bytes(ThreadContext context, IRubyObject self, IRubyObject n) {
-        return RubyString.newStringNoCopy(context.runtime, nextBytes(context, n));
+        return RubySecureRandom.random_bytes(context, self, n);
     }
 
-    @JRubyMethod(meta = true)
+    @Deprecated
     public static IRubyObject hex(ThreadContext context, IRubyObject self) {
-        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.twosComplementToHexBytes(nextBytes(context, 16), false));
+        return RubySecureRandom.hex(context, self);
     }
 
-    @JRubyMethod(meta = true)
+    @Deprecated
     public static IRubyObject hex(ThreadContext context, IRubyObject self, IRubyObject n) {
-        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.twosComplementToHexBytes(nextBytes(context, n), false));
+        return RubySecureRandom.hex(context, self, n);
     }
 
-    @JRubyMethod(meta = true)
+    @Deprecated
     public static IRubyObject uuid(ThreadContext context, IRubyObject self) {
-        return RubyString.newStringNoCopy(context.runtime, ConvertBytes.bytesToUUIDBytes(nextBytes(context, 16), false));
+        return RubySecureRandom.hex(context, self);
     }
 
-    private static byte[] nextBytes(ThreadContext context, IRubyObject n) {
-        int size = n.isNil() ? 16 : (int) n.convertToInteger().getLongValue();
-
-        return nextBytes(context, size);
-    }
-
-    private static byte[] nextBytes(ThreadContext context, int size) {
-        if (size < 0) throw context.runtime.newArgumentError("negative argument: " + size);
-
-        byte[] bytes = new byte[size];
-        context.getSecureRandom().nextBytes(bytes);
-
-        return bytes;
-    }
 }
